requirements.txt
fastapi
uvicorn[standard]
python-multipart
twilio
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
requests
asyncpg==0.29.0
python-dotenv